update adoptable-dogs internal page with external petfinder link
petfinder has finally deprecated the old API v1 that we were relying on
to retrieve the petfinder animal data. the new API v2 truncates the
`description` field at 100 characters which doesn't fit our current
page design. as a workaround, we are going to link externally to the
petfinder site so the end user has a usable experience.