Replace custom `<AdoptableDogs/>` component with Petfinder widget (#638)
* removing custom <AdoptableDogs/> code and replacing with Petfinder widget

* lint fix

* `pnpm wranger types` update worker-configuration.d.ts

* Fix Petfinder widget CORS issue

Was previously loading the Petfinder script in a way that
triggers **CORS**. That only happens if the browser treats
it as a **module** (or you try to `import`/`fetch` it).
Cross-origin **module** scripts require the other site to
send `Access-Control-Allow-Origin`, and Petfinder’s file
doesn’t.

FIX:
Load it as a **classic** script so the request is
`mode: "no-cors"`:

* add 20px spacing at the bottom of the petfinder widget

* use updated petfinder widget syntax